Create CGPA Calculator
This program allows user to calculate his/her CGPA. The user has to enter the details such as no. of courses taken, grades earned in each  course. The program also calculate total credits, total grade points achieved, determine SGPA for both semesters. On the basis of all this data, it can generate CGPA of the student.